+++ tsup.config.ts
@@ -20,7 +20,8 @@ export default defineConfig({
     testimonial: 'src/components/testimonial/index.ts',
     alert: 'src/components/alert/index.ts',
     avatar: 'src/components/avatar/index.ts',
-    spinner: 'src/components/spinner/index.ts'
+    spinner: 'src/components/spinner/index.ts',
+    tag: 'src/components/tag/index.ts'
   },
   format: ['esm', 'cjs'],
   sourcemap: true,